Fix steam version and button leads to last open among us

+++ AmongUsCapture/Memory/Structs/SteamWinningPlayerData.cs
@@ -11,7 +11,7 @@ namespace AmongUsCapture.Memory.Structs
     public struct SteamWinningPlayerData : WinningPlayerData
     {
         [FieldOffset(0x8)]
-        public long Name;
+        public uint Name;
         [FieldOffset(0xC)]
         public bool IsDead;
         [FieldOffset(0xD)]
